+++ crates/hyperqueue/src/client/commands/submit/mod.rs
@@ -4,6 +4,6 @@ pub mod directives;
 mod jobfile;
 
 pub use command::SubmitJobConfOpts;
-pub use command::{resubmit_computation, submit_computation, JobResubmitOpts, JobSubmitOpts};
+pub use command::{submit_computation, JobSubmitOpts};
 
 pub use jobfile::{submit_computation_from_job_file, JobSubmitFileOpts};