+++ crates/pyhq/python/hyperqueue/job.py
@@ -31,7 +31,11 @@ class Job:
         self.tasks: List[Task] = []
         self.task_map: Dict[TaskId, Task] = {}
         self.max_fails = max_fails
-        self.default_workdir = Path(default_workdir).resolve() if default_workdir is not None else default_workdir
+        self.default_workdir = (
+            Path(default_workdir).resolve()
+            if default_workdir is not None
+            else default_workdir
+        )
         self.default_env = default_env or {}
 
     def task_by_id(self, id: TaskId) -> Optional[Task]: