+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Breaking change
 
+* Mechanism for resubmitting tasks was changed. Command `resubmit` was removed, see the manual for replacement.
+
 * The output format of the `job info` command with JSON output mode has been changed. Note that
 the JSON output mode is still unstable.